+++ extras/actionbarcompat/src/com/manuelpeinado/multichoiceadapter/extras/actionbarcompat/MultiChoiceBaseAdapter.java
@@ -18,7 +18,6 @@ package com.manuelpeinado.multichoiceadapter.extras.actionbarcompat;
 import java.util.Set;
 
 import android.content.Context;
-import android.content.res.Resources;
 import android.os.Bundle;
 import android.support.v7.view.ActionMode;
 import android.view.View;
@@ -220,8 +219,7 @@ public abstract class MultiChoiceBaseAdapter extends BaseAdapter implements Acti
 
     @Override
     public String getActionModeTitle(int count) {
-        Resources res = getContext().getResources();
-        return res.getQuantityString(R.plurals.selected_items, count, count);
+		return helper.getActionModeTitle(count);
     }
 
     //